Move geometry to its own primary file
