+++ meerk40t/core/elements/elements.py
@@ -40,6 +40,7 @@ def plugin(kernel, lifecycle=None):
             placements,
             render,
             shapes,
+            geometry,
             trace,
             tree_commands,
             undo_redo,
@@ -54,6 +55,7 @@ def plugin(kernel, lifecycle=None):
             wordlist.plugin,
             materials.plugin,
             shapes.plugin,
+            geometry.plugin,
             tree_commands.plugin,
             undo_redo.plugin,
             clipboard.plugin,
